Compatibility with VN ROM V14.1 (quick-fix only)
Fixing issue #52 - In this ROM, the text color will fall back to the Sky
Blue default color. Further investigation of how the ROM uses a
LinearLayout and how to get the child TextView from inside it will need
to be done to fix this correctly and permanently. For now, I'll just
place a workaround to fall back to if the View can't be resolved to a
TextView.